# backend/services/ai-chatbot/openapi.yaml
swagger: '2.0'
info:
  title: Grassroots Football Management API
  description: API Gateway for Grassroots Football Management microservices
  version: 1.0.0
host: gfm-1ykrsotxghm8m.apigateway.grassroots-football-management.cloud.goog
...
schemes:
  - https
securityDefinitions:
  firebase:
    type: oauth2
    flow: "implicit"
    authorizationUrl: ""
    x-google-issuer: "https://securetoken.google.com/grassroots-football-management"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "grassroots-football-management"
security:
  - firebase: []
paths:
...
  /club/create-join:
    post:
      x-google-backend:
        address: https://microservice-team-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-team-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Create or join a club
      operationId: createOrJoinClub
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              clubName:
                type: string
              coachEmail:
                type: string
              ageGroups:
                type: array
                items:
                  type: string
              divisions:
                type: array
                items:
                  type: string
              county:
                type: string
      responses:
        '201':
          description: New club created
        '200':
          description: Coach added to existing club
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /club/search:
    get:
      x-google-backend:
        address: https://microservice-team-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-team-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Search for clubs
      operationId: searchClubs
      produces:
        - application/json
      parameters:
        - in: query
          name: clubName
          type: string
          description: Club name to search
          required: false
        - in: query
          name: county
          type: string
          description: County to filter by
          required: false
        - in: query
          name: ageGroup
          type: string
          description: Age group to filter by
          required: false
        - in: query
          name: division
          type: string
          description: Division to filter by
          required: false
      responses:
        '200':
          description: List of clubs
        '500':
          description: Internal server error
  /club/join-request:
    post:
      x-google-backend:
        address: https://microservice-team-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-team-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Submit a join request
      operationId: submitJoinRequest
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              playerEmail:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
      responses:
        '201':
          description: Join request submitted
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /club/requests:
    get:
      x-google-backend:
        address: https://microservice-team-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-team-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Retrieve pending join requests
      operationId: getJoinRequests
      produces:
        - application/json
      parameters:
        - in: query
          name: clubName
          type: string
          description: The club name to retrieve join requests for
          required: true
        - in: query
          name: ageGroup
          type: string
          description: The age group to filter join requests by
          required: false
        - in: query
          name: division
          type: string
          description: The division to filter join requests by
          required: false
      responses:
        '200':
          description: List of pending join requests
        '400':
          description: Missing or invalid query parameters
        '500':
          description: Internal server error
  /club/requests/approve:
    post:
      x-google-backend:
        address: https://microservice-team-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-team-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Approve a player's join request
      operationId: approveJoinRequest
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              playerEmail:
                type: string
                description: Email of the player to approve
              clubName:
                type: string
                description: The club name where the player requested to join
              ageGroup:
                type: string
                description: The age group of the player
              division:
                type: string
                description: The division of the player
      responses:
        '200':
          description: Player added to the club
        '400':
          description: Missing or invalid parameters
        '500':
          description: Internal server error
  /club/requests/reject:
    post:
      x-google-backend:
        address: https://microservice-team-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-team-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Reject a player's join request
      operationId: rejectJoinRequest
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              playerEmail:
                type: string
                description: Email of the player to reject
              clubName:
                type: string
                description: The club name where the player requested to join
              ageGroup:
                type: string
                description: The age group of the player
              division: 
                type: string
                description: The division of the player
      responses:
        '200':
          description: Join request rejected
        '400':
          description: Missing or invalid parameters
        '500':
          description: Internal server error
  /club/players:
    get:
      x-google-backend:
        address: https://microservice-team-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-team-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Retrieve players for a specific club, age group, and division
      operationId: getPlayers
      produces:
        - application/json
      parameters:
        - in: query
          name: clubName
          type: string
          description: The club name to retrieve players for
          required: true
        - in: query
          name: ageGroup
          type: string
          description: The age group to filter players by
          required: true
        - in: query
          name: division
          type: string
          description: The division to filter players by
          required: true
      responses:
        '200':
          description: List of players for the specified criteria
        '400':
          description: Missing or invalid query parameters
        '500':
          description: Internal server error


  # Schedule Management Microservice Routes
...